+++ test/actions/BackendAction-test.js
@@ -125,20 +125,14 @@ describe('BackendAction', () => {
   describe('watching', () => {
     it('should update pipelines', (done) => {
       backendAction.extensionConfigWatcher = {
-        start: () => { return backendAction.extensionConfigWatcher },
-        stop: (cb) => { cb() },
-        on: (name, cb) => { cb() }
-      }
-      backendAction.themeConfigWatcher = {
-        start: () => { return backendAction.themeConfigWatcher },
-        stop: (cb) => { cb() },
-        on: (name, cb) => { cb() }
+        start: (cb) => cb(),
+        stop: (cb) => cb(),
+        on: (name, fn) => fn()
       }
 
-      backendAction.dcClient.downloadPipelines = (appId, cb) => cb(null, [{pipeline: {id: 'testPipeline'}}])
-      backendAction.dcClient.removePipeline = (pId, aId, cb) => cb()
-      backendAction._extensionChanged = (cfg, cb = () => {}) => {}
-      backendAction._themeChanged = (cfg, cb = () => {}) => {}
+      backendAction.dcClient.downloadPipelines = (appId, trusted, cb) => cb(null, [{pipeline: {id: 'testPipeline'}}])
+      backendAction.dcClient.removePipeline = (pId, aId, trusted, cb) => cb()
+      backendAction._extensionChanged = (cfg, cb = () => {}) => cb()
 
       try {
         backendAction._startSubProcess()
@@ -163,7 +157,7 @@ describe('BackendAction', () => {
       const file = path.join(backendAction.pipelinesFolder, 'dCPlTest.json')
       assert.equal(backendAction.pipelines[file], undefined)
 
-      backendAction.dcClient.uploadPipeline = (f, aId, cb) => {
+      backendAction.dcClient.uploadPipeline = (f, aId, trusted, cb) => {
         assert.deepEqual(backendAction.pipelines[file].id, pipeline.pipeline.id)
         assert.deepEqual(f, pipeline)
         assert.equal(aId, appId)
@@ -206,7 +200,7 @@ describe('BackendAction', () => {
 
     it('should throw error if dcClient is not reachable', (done) => {
       const pipeline = {pipeline: {id: 'plFooBarline2'}}
-      backendAction.dcClient.uploadPipeline = (pl, id, cb) => cb(new Error(`error`))
+      backendAction.dcClient.uploadPipeline = (pl, id, trusted, cb) => cb(new Error('error'))
 
       const file = path.join(backendAction.pipelinesFolder, 'dCPlTest2.json')
       fsEx.writeJson(file, pipeline, (err) => {
@@ -217,11 +211,11 @@ describe('BackendAction', () => {
           done()
         })
       })
-    }).timeout(5000)
+    })
 
     it('should return if pipeline was changed', (done) => {
       const pipeline = {pipeline: {id: 'plFooBarline3'}}
-      backendAction.dcClient.uploadPipeline = (pl, id, cb) => cb()
+      backendAction.dcClient.uploadPipeline = (pl, id, trusted, cb) => cb()
 
       const file = path.join(backendAction.pipelinesFolder, 'dCPlTest3.json')
       fsEx.writeJson(file, pipeline, (err) => {
@@ -236,7 +230,7 @@ describe('BackendAction', () => {
     it('should return if pipeline was removed', (done) => {
       const pipelineId = 'plFooBarline3'
       let called = false
-      backendAction.dcClient.removePipeline = (plId, id, cb) => {
+      backendAction.dcClient.removePipeline = (plId, id, trusted, cb) => {
         assert.equal(plId, pipelineId)
         called = true
         cb()